services/AirflowService/pyproject.toml
[project]
name = "airflowservice"
version = "1.0.0"
description = "Airflow orchestration for notification-bot pipelines"
requires-python = ">=3.10"
dependencies = [
    "apache-airflow>=2.9.0",
    "pyarrow>=17.0.0",
    "redis>=5.0.0",
    "httpx>=0.27.0",
    "python-dotenv>=1.0",
]
